refactor(debug-bar): share client-build loaders between the two builds
Extract the esm-env strip, mochi-env virtual module, and .svelte.[jt]s
module loaders into clientBuildLoaders.ts, registered by both the island
client build and the standalone debug bar build so they can't drift. The
debug bar's svelte compiles now also go through mergeCompilerOptions
(framework defaults, still no user config).
